Unify provisioning, solve the dependency
