+++ cfme/tests/infrastructure/test_provisioning_dialog.py
@@ -3,10 +3,8 @@ import pytest
 import re
 from datetime import datetime, timedelta
 
-import cfme.infrastructure.provisioning
-assert cfme.infrastructure.provisioning
-from cfme.web_ui.prov_form import provisioning_form
 from cfme.infrastructure.virtual_machines import Vm, details_page
+from cfme.provisioning import provisioning_form
 from cfme.services import requests
 from cfme.web_ui import fill, flash
 from utils import testgen, version
